// src/shared/api/models/putApiCampaignsIdBody.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { PutApiCampaignsIdBodyDescription } from './putApiCampaignsIdBodyDescription';
import type { PutApiCampaignsIdBodyMessageType } from './putApiCampaignsIdBodyMessageType';
import type { PutApiCampaignsIdBodyTestPhoneNumber } from './putApiCampaignsIdBodyTestPhoneNumber';
import type { PutApiCampaignsIdBodyTestEmail } from './putApiCampaignsIdBodyTestEmail';
import type { PutApiCampaignsIdBodyBccEmail } from './putApiCampaignsIdBodyBccEmail';
import type { PutApiCampaignsIdBodyConditionsItem } from './putApiCampaignsIdBodyConditionsItem';
import type { PutApiCampaignsIdBodyContent } from './putApiCampaignsIdBodyContent';
import type { PutApiCampaignsIdBodySchedule } from './putApiCampaignsIdBodySchedule';

export type PutApiCampaignsIdBody = {
    /** @minLength 1 */
    name?: string;
    description?: PutApiCampaignsIdBodyDescription;
    messageType?: PutApiCampaignsIdBodyMessageType;
    /** Test phone number for campaign testing */
    testPhoneNumber?: PutApiCampaignsIdBodyTestPhoneNumber;
    /** Test email for campaign testing */
    testEmail?: PutApiCampaignsIdBodyTestEmail;
    /** BCC email address for campaign */
    bccEmail?: PutApiCampaignsIdBodyBccEmail;
    /** Conditions for targeting customers. Use either template-based (condition_id + field_value + operator + seq_id + render_id) or custom conditions (entity + field + operator + field_value) */
    conditions?: PutApiCampaignsIdBodyConditionsItem[];
    content?: PutApiCampaignsIdBodyContent;
    schedule?: PutApiCampaignsIdBodySchedule;
};